src/JobStats: move stats component in separate file
